Compile namespace assignment checks earlier
This fixes a bug that existed in the compiler because of the previous
assumption that namespacees would only ever be used when assigning to a
single variable. Single it is not allowed to be properly used asa n
expression, that assumption is no longer accurate and as such we needed
to update the compiler to check that the namespace assignments are in
fact pointing to a namespace earlier. Otherwise we ran into a syntax
error that was coming from the assignment check being done within the
tuple.